Get code completion suggestions working (mostly)
I've updated the code completions to actually work for the most part.  I needed to override a lot of the functionality around the completion UI element.

To support this, I also needed to fork gxui (and use my fork) for the time being - it seems difficult to handle automatic code features without access to the *TextBoxController field of mixins.TextBox.