Big restructure, plus navigation pane
I felt that things were getting a little hard to understand, so I drew out a picture of the editor's UI structure as it
stood, then restructured the code to match the structure of the UI.  This takes care of cleaning up a lot of that, as well
as making the command box much more powerful.

A navigation pane has been added.  It currently includes buttons for a "projects" UI and a "folders" UI (typically for
displaying the folder tree of a project).  I've got a config directory included, as well, which stores projects that have
been added.

Also, I've moved from codegangsta/cli to spf13/cobra for a time.  I'm curious how the two compare.