Enable actual projects selection
The projects pane now directly affects the files pane.  When selecting a project, the files pane will switch to only showing the files/directories inside of a project.

Some small hurdles made it hard to make the files pane a focusable, so I've swapped to just making panes a gxui.Control for the moment.  I'll be switching back to focusable soon, though.