Fix weird duplicate key bug
The call to KeyPress on the window's OnKeyUp was intended to take care of
the case where there was no focus.  I didn't take into account the fact that
OnKeyUp doesn't have a way to consume an event, though, so most commands
were called twice for each key stroke.  This was especially bad with the
cut command and made the issue obvious.