Added file/edit menu
The commander can now act as a generic container.  With a linear layout, elements aren't allowed to overlap
each other, which made it difficult to overlay the main UI with a dropdown menu.  With these changes, the
dropdown menu idea works, and I've added menus for the commands to live under.