Splitting and focusing within a split now works
Yet to do: splitter layouts don't seem to play well with each other (e.g. replacing a tabbed editor in a split layout
with a split layout of a different orientation).  This seems like it might be a bug related to splitter layouts, so
the fix will likely go in gxui.