Restore the last position set by user input
To prevent the editor from scrolling back to the top of a file on save, we need to keep
a record of the last selection and scroll positions that were set by **user input**.
Then, when the editor text is updated, we always restore those positions.  This prevents
the truncation of the file from forcing a scroll to the top of the file that we can't
recover from.