Update syntax package
The first few tests in the syntax package exposed issues around duplicate entries in
the CodeSyntaxLayers returned by syntax.Layers.  When trying to fix that, more issues
were exposed related to ordering of the values in the result, as well as unnecessarily
complicated structure of data (we only really need one CodeSyntaxLayer for each color).

All of this resulted in a pretty large refactor of the syntax highlighting logic.