Update all configurations to use viper
Project and vidar settings are both now handled by vidar.  There is logic in place to move from the old
yaml files to the new files.

Since it's been quite a while since I changed the name of the key bindings for home/end, I also removed
the logic to convert from the old to the new names there.