Draw a resize cursor when mousing over the splitter bar
Make our own splitter bar type which contains MouseEnter and MouseExit methods that make use of
the viewport's new SetCursor method to make it obvious when the cursor is in a location that
can resize the view.

While we're at it, get rid of those Init methods.  I dislike them.