Work in progress rearchitecting nav bar
While trying to make sure the navigation bar updates with filesystem changes, I found out that
the mixins.Tree in gxui automatically collapses all elements whenever the data is changed in
any node.  This isn't ideal for a pane that's used to monitor files.

The ideal solution seems to be to build a new tree architecture using LinearLayouts filled with
buttons.  I have the first pass of this work done, but it *only* deals with directory structure
at the moment.  I'll need to update the table of contents tree similarly, and probably the
projects list, too.