Allow hooks to bind to opening a file
All file-specific commands now bind when a file is opened and an editor gains
focus.  This allows us to conditionally add functionality depending on what
has focus, or what type of file is open.  For instance, we can only add a
'goimports' command if a file ending in '.go' is open.