More plugin usage, esp. additional thread safety
A lot of old code relied on the UI thread to do a big chunk of work.  As
plugins are supported more fully, we want to reduce reliance on the UI
thread to allow the editor to continue operating as normally as possible
when work is happening in plugins.  This requires more locks and less
reliance on driver.Call.