Syntax highlighting cleanup - rainbow fixes, etc
Rainbow highlighting wasn't working right because the syntax layers weren't sorted before getting to the editor's logic for turning syntax layers into colors.  We now sort the slice before parsing it.

In addition, there was some unnecessary BS in the rainbow code, where it would request a random highlight even if that was unneeded.  The logic could probably still be improved, but it's working fine for the moment.

The final fix here was to avoid including a pointer to the rainbow in the theme.  This means that a rainbow will be copied upon passing to new editors, so each editor gets its own rainbow type.