Bindings updates to make hooks more consistent
HookedCommand types were previously required to implement a Clone method, which was
used to clone them when a new editor was opened if they were already bound.  The
problem with this method was that the hooks that were being bound anew were *not*
cloned, resulting in multiple calls to Bind, re-binding already-bound hooks to the
HookedCommand.

This update makes it so that only Op types (types which execute actions) can bind hooks,
and requires that they return a new copy of themselves with the hook bound for each
hook that binds to them.  While this is a bit of a memory churn, it definitely feels
cleaner and safer (much less chance of binding a hook multiple times), and it happens infrequently enough that I think it's a reasonable choice until such a time as
the editor starts feeling sluggish when switching tabs and whatnot.