Update the syntax highlighting plugin to adjust spans as edits happen
It will still fully re-parse the syntax tree on every edit, but because
that happens in a separate goroutine, it was causing flickers as the
text changed without the spans changing.  With the Apply hook, we adjust
the current layers' positions while we wait on the full re-parse to finish.