Solve menu mapping bug for commands that allow hooks to bind to them
Commands that bind hooks will often change their hash with bound hooks, since the slice
of hooks will be different with each hook that binds to them.  This was hiding command
mappings in the menu, since the map was using the hash of the command, instead of the
command name.