Override AddPanelAt in TabbedEditor
Dragging panels between tabbed editors was causing the new tabbed editor
to add the panel as a child without also storing that editor in its map
of available editors.  This meant that the next time something tried to
focus the editor, it would open a second copy of that tab.

This fixes that issue.

Resolves #107