plugin/gocode: actually close when there are no results to apply
Prior to this, there were cases where gocode would have an empty list and try to apply a nonexistent code suggestion.  We've prevented these types of issues from crashing the editor, but we didn't fix the plugin (partly because we wanted to use the plugin to verify the fixes in the editor).  These are now fixed.