fsw: use poller for windows
At least in Wine, the windows fsnotify code just doesn't seem to be working
very well.  It seems to me to make sense for windows and darwin to share
the polling code.  It helps me test the windows version - even if my problems
are just related to Wine, that seems worth it.