plugin/gosyntax: lock access to the syntax highlighter
As per #147, we were seeing occasional concurrent map writes and reads.  This is the quick, obvious
fix for the issue, although we should take another look after the fix is confirmed to see if we
really need the _same_ syntax parser in both goroutines.  On the one hand, we probably don't want
to assume that all of the code will be re-parsed from scratch; on the other hand, we're not passing
along any information about which spans could have changed.

Resolves #147