settings: drop font and config file dependencies
We're dropping tmc/fonts as a dependency mostly because it's a very small dependency and we uncovered
a bug in it.  'A little bit of copy/paste is better than a little bit of dependency' seems like a
sane reason for us to drop it and write our own.  It also gives us some flexibility to look for fonts
in the XDG dirs.

We're dropping viper because we submitted github.com/spf13/viper#271 ages ago, but it never
got a response.  I don't want to keep our workaround code any more.

We've switched to OpenPeeDeeP/xdg for our xdg code mostly because it seems to be more maintained and has a
clear README.