settings/config: fix json decoding
json.Decoder appears to ... either be buggy or have a _questionable_ optimization.  It does not
read until EOF, instead reading until it believes the file is completely read.  Since our tests
expect the Decoder to read until EOF (and since I frankly consider json's behavior a bug), we're
now doing the byte reading ourselves with a call to ioutil.ReadAll and passing the byte buffer
to json.Unmarshal instead of using json.Decoder.