command/history: more cleanup - make sure history.go is not importing unsafe or atomic
In order to more easily prove correctness in the new atomic/unsafe logic in history,
we moved all of the methods which were making use of atomic/unsafe into a separate
file which encompases mostly just getters and setters.  This makes it much easier
to prove that history.go is not using atomic/unsafe directly and instead relying on
those methods.