.travis.yml: remove go tip
I added tip as an allowed failure so that we'd get something of a heads up if anything broke with
in-development versions of go.  Unfortunately, the 'allow failure' setting in travis does not
allow the build to complete until the allowed failures are finished - making this more counter-
productive than it's worth.