modules: fix up vidar for working on modules
Go modules are in heavy use, for better or for worse.  We need to support them.
The project type had to change in the settings in order to track a host of
environment variables, especially since the GOPATH is no longer required.  This
forced changes to several plugins (e.g. goimports and godef), as well.